Add `setupAppModule` & `setupLibraryModule`
