+++ buildSrc/build.gradle.kts
@@ -1,0 +1,10 @@
+plugins {
+  `kotlin-dsl`
+}
+
+apply("../gradle/extra.gradle.kts")
+
+dependencies {
+  implementation(rootProject.extra["androidPlugin"].toString())
+  implementation(rootProject.extra["kotlinPlugin"].toString())
+}
